requirements.txt
streamlit
pymongo
python-dotenv
requests
transformers
PyJWT==2.8.0 